chat1/requirements.txt
aioredis==1.2.0
asgiref==3.7.2
astroid==2.0.4
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2018.10.15
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==3.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
daphne==4.0.0
defusedxml==0.5.0
Django==4.2.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
drf-yasg2==1.19.4
gunicorn==19.7.1
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==22.10.0
inflection==0.5.1
isort==4.3.4
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.3.1
loguru==0.7.2
MarkupSafe==2.1.3
mccabe==0.6.1
msgpack==1.0.6
oauthlib==2.1.0
openapi-codec==1.3.2
packaging==23.1
pipdeptree==2.13.0
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyHamcrest==1.9.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
python3-openid==3.1.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.0.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
service-identity==23.1.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
twisted==23.8.0
txaio==23.1.1
typing-extensions==4.8.0
uritemplate==4.1.1
urllib3==1.25.2
wrapt==1.10.11
zope.interface==6.0
